requirements.txt
geopandas
dask_geopandas
pyogrio
matplotlib
contextily
